requirements.txt
ytmusicapi
python-dotenv
google-genai